services/embed-loader/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=3.20.0
numpy>=1.24.0
